feat(beamer-injected): add convention, refactor to convention
