+++ docs/1_after_pr.md
@@ -1,8 +1,8 @@
-# After PR
+# 🧙 After PR
 
 _made with love by [Mau](https://github.com/maurodibert) and [Martin](https://github.com/mal2tin)  for the_ **Monkslab Team** ❤️
 
-## 🧙 Wise, targeted and valuable solutions
+## Wise, targeted and valuable solutions
 
 - [x] 1. git checkout main
 - [x] 2. git pull to main branch
