+++ docs/3_before_pr.md
@@ -1,8 +1,8 @@
-# Before PR
+# 🚀  Before PR
 
 _made with love by [Mau](https://github.com/maurodibert) and [Martin](https://github.com/mal2tin)  for the_ **Monkslab Team** ❤️
 
-## 📲 Get MAIN code, and align
+## Get MAIN code, and align
 
 - [x] 1. run all the test: flutter test --coverage and remove unnecessary files with: lcov --remove coverage/lcov.info '**/*.g.dart' '**/*.freezed.dart' -o coverage/lcov.info
 - [ ] 2. test in iOS, android and web
@@ -21,7 +21,7 @@ _made with love by [Mau](https://github.com/maurodibert) and [Martin](https://gi
 - [ ] 11. merge into your branch
 - [ ] 12. repeat from step 1 until code is cuter than eating pancakes on a rainy day
 
-## 🚀 Updates
+## Updates
 
 - [ ] 1. upgrade with [Semantic Versioning](https://semver.org/) in pubspec.yaml
   - **major** _version when you make incompatible API changes_
@@ -36,7 +36,7 @@ _made with love by [Mau](https://github.com/maurodibert) and [Martin](https://gi
 - [ ] 5. update **screenshots** in [readme](../README.md)
 - [ ] 6. update **tests** if there was a refactor
 
-## 🎯 Final Stage
+## Final Stage
 
 - [ ] 1. take the screen record you did, save it with the name of the feature and the version (for ex: url_shortener_0.1.0), and add it to [project readme](../README.md) _if adds something new and useful_
 - [ ] 2. auto-review PR locally
