+++ docs/tech_debt.md
@@ -1,4 +1,4 @@
-# Tech debt
+# ☺️ Tech debt
 
 A place for you to put good ideas, error to avoid or anything interesting for this and next projects to tackle.
 _made with love by [Mau](https://github.com/maurodibert) and [Martin](https://github.com/mal2tin) for the_ **Monkslab Team** ❤️
