+++ lib/src/constants/index.dart
@@ -1,5 +1,5 @@
 export './courses/index.dart';
 export '../ui/colors.dart';
 export '../ui/text_styles.dart';
-export 'app_assets.dart';
-export 'app_models.dart';
+export 'assets.dart';
+export 'models.dart';
