+++ lib/src/ui/widgets/profile_photo.dart
