+++ lib/src/ui/widgets/socials.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:the_monkslab_web/src/constants/app_models.dart';
+import 'package:the_monkslab_web/src/constants/models.dart';
 import 'package:the_monkslab_web/src/ui/sizes.dart';
 
 import 'index.dart';
